+++ src/core/types/enums/ModelsLayout.enum.ts
@@ -1,0 +1,4 @@
+export enum ModelsLayout {
+    BUNDLE = 'bundle',
+    PER_FILE = 'per-file',
+}